Reads and modifies the SWAT projects' filo.cio according to provided inputs

Reads and modifies the SWAT projects' filo.cio according to provided inputs

Usage

setup_swat2012(
  project_path,
  output,
  start_date,
  end_date,
  output_interval,
  years_skip,
  rch_out_var,
  sub_out_var,
  hru_out_var,
  hru_out_nr
)

Arguments

project_path

Path to the SWAT project folder (i.e. TxtInOut)

start_date

Start date of the SWAT simulation. Provided as character string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format project are located

end_date

End date of the SWAT simulation. Provided as character string in any ymd format (e.g. 'yyyy-mm-dd') or in Date format project are located

output_interval

Time interval in which the SWAT model outputs are written. Provided either as character string ("d" for daily, "m" for monthly, or "y" for yearly) or as SWAT input values (0 for monthly, 1 for daily, 2 for yearly).

years_skip

Integer value that provides the numbe of years to be skipped during writing the SWAT model outputs

rch_out_var

Numeric vector of maximum length = 20 for customized output of reach variables.For output codes see SWAT I/O Documentation p.77ff.

sub_out_var

Numeric vector of maximum length = 15 for customized output of subbasin variables.For output codes see SWAT I/O Documentation p.78ff.

hru_out_var

Numeric vector of maximum length = 20 for customized output of HRU variables.For output codes see SWAT I/O Documentation p.79ff.

hru_out_nr

Numeric vector of maximum length = 20 for providing the HRU numbers for which the HRU variables are written. Optional if hru_out_nr = 'all', HRU variables are written for all HRU (caution, very large output files possible!)